update to unrar 3.7.7

+++ lib/UnrarXLib/file.cpp
@@ -300,10 +300,12 @@ void File::Flush()
 
 bool File::Delete()
 {
-  /*if (HandleType!=FILE_HANDLENORMAL || !AllowDelete)
+  /*if (HandleType!=FILE_HANDLENORMAL)
     return(false);
-  if (hFile!=BAD_HANDLE)
-    Close();
+   if (hFile!=BAD_HANDLE)
+     Close();
+  if (!AllowDelete)
+    return(false);
   return(DelFile(FileName,FileNameW));*/
   return kodi::vfs::DeleteFile(FileName);
 }
